refactor: move JwtProcessingError and JwtServiceInitError to jwt/error.rs
